Add low quality png for lazy loaded img. Fix lazyload bug. Refactor to prepare for future changes.
