+++ src/ImageProcessor.Web.Episerver/Extensions/Picture/Enums.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+//TODO: namespace should be ImageProcessor.Web.Episerver.Picture, but wait for other breaking changes(?)
+namespace ImageProcessor.Web.Episerver
+{
+	public enum LazyLoadType
+	{
+		None,
+		Regular,
+		Progressive
+	}
+}
